Another fixes

+++ GameServer_7/gameserver/ClientCommand.go
@@ -12,11 +12,11 @@ const (
 type ClientCommand struct {
 	ID          uint32  `json:"id"`
 	Type        uint8   `json:"type"`
-	X           float32 `json:"x"`
-	Y           float32 `json:"y"`
-	VX          float32 `json:"vx"`
-	VY          float32 `json:"vy"`
-	Duration    float32 `json:"duration"`
+	X           float64 `json:"x"`
+	Y           float64 `json:"y"`
+	VX          float64 `json:"vx"`
+	VY          float64 `json:"vy"`
+	Duration    float64 `json:"duration"`
 	VisualState uint8   `json:"visualState"`
 	AnimName    string  `json:"animName"`
 }
